Skybox implemented
Skybox is fully functional and looks good.

+++ src/Input/InputManager.cpp
@@ -27,10 +27,13 @@ bool InputManager::IsKeyPressed(unsigned char key)
 
 void InputManager::MouseMotion(int x, int y)
 {	
+	/* Calculates the distance the cursor has moved from the centre of the screen */
 	float screen_width = glutGet(GLUT_SCREEN_WIDTH);
 	float screen_height = glutGet(GLUT_SCREEN_HEIGHT);
-	float middle_screen_height = screen_height / 2;
-	float middle_screen_width = screen_width / 2;
+	float middle_screen_height = (int) (screen_height / 2);
+	float middle_screen_width = (int) (screen_width / 2);
+
+	//std::cout << "middle height " << middle_screen_height << ", middle width " << middle_screen_width << std::endl;
 	this->mouseX = x - middle_screen_width;
 	this->mouseY = y - middle_screen_height;
 }
